Add fields in Applicant Page

+++ src/main/java/Application/DataModel/ProfRepo.java
@@ -2,6 +2,6 @@ package Application.DataModel;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProfRepo extends CrudRepository<Professor, Integer> {
+public interface ProfRepo extends CrudRepository<Professor, Long> {
 
 }
